// src/types/api/household.ts
/**
 * LifeMapWebAPI
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: v1
 *
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import {Photo} from "~/types/api/photo";
import {Person} from "~/types/api/person";
import {User} from "~/types/api/user";
import { DirectionTask } from "./directionTask";


export interface Household { 
    HouseholdID: number;
    HouseholdName?: string;
    PrimaryPerson1ID?: number;
    PrimaryPerson2ID?: number;
    MaritalStatusID?: number;
    MarriageDate?: Date;
    ReligiousAffiliation?: string;
    PhoneNumberListID?: number;
    AddressListID?: number;
    CommentSetID?: number;
    PhotoAlbumID?: number;
    PhotoURL?: string;
    CreatedBy ?: string;
    LiveData ?: boolean;
    CouplePhotoURL ?: string;
    Photo?: Photo;
    CouplePhotoObject?:Photo;
    Photos?: Photo[];
    Persons?: Person[];
    DirectionProgress?: DirectionTask[];
    Users?: User[];
    HouseholdChecklist?:any;
    AccessType: number;
    FamilyName?: string,
    FamilyPhoto?: string
}